Refactor UDP thread and fix a bug in reading UDP packets
