+++ utils/src/elv_channel.c
@@ -8,6 +8,7 @@
 #include <sys/time.h>
 
 #include "elv_channel.h"
+#include "elv_log.h"
 
 struct elv_channel_t
 {
@@ -128,6 +129,7 @@ elv_channel_timed_receive(
         rc = pthread_cond_timedwait(&channel->_cond_send, &channel->_mutex, &ts);
         /* ETIMEDOUT is not a real error */
         if (rc != 0) {
+            elv_log("XXX pthread_cond_timedwait rc=%d, count=%d, front=%d, rear=%d", rc, channel->_count, channel->_front, channel->_rear);
             pthread_mutex_unlock(&channel->_mutex);
             return rc;
         }
@@ -141,10 +143,10 @@ elv_channel_timed_receive(
     channel->_count--;
     msg = channel->_items[channel->_front];
     channel->_front = (channel->_front+1) % channel->_capacity;
+    *rcvdmsg = msg;
     pthread_cond_signal(&channel->_cond_recv);
     pthread_mutex_unlock(&channel->_mutex);
 
-    *rcvdmsg = msg;
     return 0;
 }
 
@@ -175,6 +177,7 @@ elv_channel_close(
     pthread_cond_signal(&channel->_cond_recv);
     pthread_cond_signal(&channel->_cond_send);
 
+    elv_log("XXX elv_channel_close count=%d, front=%d, rear=%d", channel->_count, channel->_front, channel->_rear);
     /* Purge the channel if the flag is set */
     while (purge_channel && channel->_count > 0) {
         channel->_count--;
